refactor(theming): break themes→_base→presets cyclic imports + SW event.origin check
CodeQL flagged 872 py/unsafe-cyclic-import alerts across ~55 theme
files. Root cause: themes/_base.py imports from ..presets at module
level, and presets.py imports from .themes.X at module level — a real
cycle that works today only because ColorScale etc. are defined in
presets.py BEFORE line 214 where theme imports start.

Extracted shared dataclasses and instances into dependency-free
modules:

- theming/_types.py: 14 dataclass types (ColorScale, ThemeTokens,
  ThemePreset, SurfaceTreatment, TypographyStyle, LayoutStyle,
  SurfaceStyle, IconStyle, AnimationStyle, InteractionStyle,
  DesignSystem, PatternStyle, IllustrationStyle, ThemePack).
  Stdlib-only imports.
- theming/_constants.py: ~60 shared style instances (PATTERN_*,
  ILLUST_*, ICON_*, ANIM_*, INTERACT_*) at both DS and pack levels.
  Depends only on _types.

themes/_base.py now imports types from _types.py and instances from
_constants.py — bypassing the cyclic path through presets.py /
theme_packs.py. presets.py and theme_packs.py import from the new
modules and re-export for backward compat. Zero theme-author impact.

Also unified two conflicting InteractionStyle class definitions in
theme_packs.py (the narrower DS version was silently shadowed by the
wider pack version at runtime; all INTERACT_* instances relied on the
wider version). Preserved the runtime DS-vs-pack distinction for
INTERACT_MINIMAL / INTERACT_PLAYFUL via internal _INTERACT_*_DS
aliases.

Also added an explicit event.origin === self.location.origin check to
the service-worker message handler to satisfy CodeQL's
js/missing-origin-check rule (alert #2170 regression after PR #925's
source+scope check).

Expected alert closure: 872 py/unsafe-cyclic-import + 1
js/missing-origin-check = 873 alerts (pending CodeQL rescan).

7 new regression cases in test_theming_imports_backcompat.py:
- presets re-exports ColorScale/ThemeTokens/ThemePreset/SurfaceTreatment
- theme_packs re-exports all 10 types + 7 instance samples
- _base re-exports are the SAME object identity (is-comparison)
- vercel per-theme file still produces a PRESET/PACK/DESIGN_SYSTEM
- lazy theme-pack registry still populates (71 packs + 73 design systems)
- InteractionStyle DS-vs-pack distinction preserved

Full Python suite: 4429 passed, 3 failed (all pre-existing on origin/main
and unrelated to this refactor — API dispatch + observability error-
message content assertions).
