+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- **Break `themes → _base → presets/theme_packs` cyclic import (873 CodeQL alerts) + add explicit `event.origin` check to service worker `message` handler** — CodeQL's `py/unsafe-cyclic-import` rule flagged 872 alerts across the theming subsystem: `themes/_base.py` imported dataclasses + shared style instances from `..presets` and `..theme_packs`, and those two modules re-imported each theme file under `.themes.*` at module load — a real cycle that happened to work only because `ColorScale` / `ThemeTokens` / etc. were defined earlier in `presets.py` than the theme imports. Extracted the pure data into two new dependency-free modules: `python/djust/theming/_types.py` (14 dataclass types: `ColorScale`, `ThemeTokens`, `SurfaceTreatment`, `ThemePreset`, `TypographyStyle`, `LayoutStyle`, `SurfaceStyle`, `IconStyle`, `AnimationStyle`, `InteractionStyle`, `DesignSystem`, `PatternStyle`, `IllustrationStyle`, `ThemePack` — stdlib imports only) and `python/djust/theming/_constants.py` (~60 shared style instances — `PATTERN_*`, `ILLUST_*`, `ICON_*`, `ANIM_*`, `INTERACT_*` at both the design-system and pack levels; depends only on `_types`). `themes/_base.py` now imports from those two modules, bypassing the cycle; `presets.py` and `theme_packs.py` import from the same new modules and re-export every type and instance under `__all__` for full backward compat (no theme author touches any import site). Also resolved the pre-existing shadow between two `InteractionStyle` class definitions (the narrow DS-level `InteractionStyle` at `theme_packs.py:150` was silently shadowed by the wider pack-level one at `:1374` — all `INTERACT_*` module-level instances relied on fields only the wider class had; unified on the superset definition in `_types.py`) and the `INTERACT_MINIMAL` / `INTERACT_PLAYFUL` name collision between the DS-level and pack-level bindings (kept the distinct runtime bindings via `_INTERACT_MINIMAL_DS` / `_INTERACT_PLAYFUL_DS`). Also tightened the service-worker `message` handler in `python/djust/static/djust/service-worker.js` with an explicit `event.origin !== self.location.origin` early return at the top of the listener, satisfying CodeQL's `js/missing-origin-check` rule (alert #2170 — follow-up to the source+scope check shipped in #925). 7 regression cases in `python/djust/tests/test_theming_imports_backcompat.py` cover: presets/theme_packs type exports still importable, shared instance exports still importable, `_base` re-exports identical object identity to `presets` / `theme_packs`, per-theme files (vercel used as smoke) still construct a full triple, lazy theme-pack registry still populates 71 packs + 73 design systems, and the DS-vs-pack `InteractionStyle` distinction for `minimal` / `playful` is preserved (DS `link_hover="underline"`, pack `button_click="ripple"` — both bindings round-trip). Expected alert closure: 872 × `py/unsafe-cyclic-import` + 1 × `js/missing-origin-check` = 873. (`python/djust/theming/_types.py`, `python/djust/theming/_constants.py`, `python/djust/theming/presets.py`, `python/djust/theming/theme_packs.py`, `python/djust/theming/themes/_base.py`, `python/djust/static/djust/service-worker.js`)
+
 - **Dead conditional in `djust/theming/templatetags/theme_form_tags.py`** — the label-visibility check at line 88 had `isinstance(field.widget, template.library.InvalidTemplateLibrary if False else type(None))`. The `if False else type(None)` ternary always evaluated to `type(None)`, making the first operand unreachable dead code (CodeQL `py/constant-conditional-expression`). Dropped the dead branch; the isinstance check is now `isinstance(field.widget, type(None))` with a comment explaining the intent.
 
 - **Close 21 `py/undefined-export` CodeQL alerts** — `djust/auth/__init__.py` and `djust/tenants/__init__.py` use a `__getattr__`-based lazy-import dispatcher to defer Django-ORM-dependent imports. CodeQL's static analysis doesn't recognize this pattern; names declared in `__all__` but only resolved via `__getattr__` were flagged. Added a `TYPE_CHECKING` block to each `__init__.py` with eager import statements gated behind `if TYPE_CHECKING:` — the imports execute only under static analysis (mypy, CodeQL, IDEs), never at runtime. The lazy-import runtime behavior is unchanged. New `python/djust/tests/test_lazy_import_resolution.py` (47 parameterized cases) regression-tests that every `__all__` entry resolves.
